chore: move tests to new dashboards-bar dir amd outdated snapshots
