+++ src/components/DashboardsBarNew/__tests__/Chip.spec.js
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react'
 import { createMemoryHistory } from 'history'
 import React from 'react'
 import { Router } from 'react-router-dom'
-import Chip from '../Chip.js'
+import Chip from '../../DashboardsBar/Chip.js'
 
 /* eslint-disable react/prop-types */
 jest.mock('@dhis2/ui', () => {
