+++ src/components/DashboardsBarNew/__tests__/ClearButton.spec.js
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react'
 import React from 'react'
-import ClearButton from '../ClearButton.js'
+import ClearButton from '../../DashboardsBar/ClearButton.js'
 
 test('ClearButton renders a button', () => {
     const { container } = render(<ClearButton onClear={jest.fn()} />)
