+++ src/components/DashboardsBarNew/__tests__/Filter.spec.js
@@ -2,8 +2,8 @@ import { render } from '@testing-library/react'
 import React from 'react'
 import { Provider } from 'react-redux'
 import configureMockStore from 'redux-mock-store'
-import WindowDimensionsProvider from '../../../components/WindowDimensionsProvider.js'
-import Filter from '../Filter.js'
+import WindowDimensionsProvider from '../../WindowDimensionsProvider.js'
+import Filter from '../../DashboardsBar/Filter.js'
 
 const mockStore = configureMockStore()
